settings screen now accepts empty values.
Previously, empty values are interpreted as Null, which cannot be cast to String or Int. Introduce helpers to insert default values (i.e. ''  for String or 0 for int) in cases where the user set a setting to 'null'. TODO: better handling of when height gets removed. We should probably call a meteor method rather than send updateHeight with 0.